requirements.txt
pandas
numpy
scikit-learn
joblib
streamlit
google-generativeai
matplotlib
seaborn
python-dotenv
